+++ lib/sbmt/kafka_consumer/config/kafka.rb
@@ -7,12 +7,26 @@ class Sbmt::KafkaConsumer::Config::Kafka < Dry::Struct
   SERVERS_REGEXP = /^[a-z\d.\-:]+(,[a-z\d.\-:]+)*$/.freeze
 
   attribute :servers, Sbmt::KafkaConsumer::Types::String.constrained(format: SERVERS_REGEXP)
-  attribute? :kafka_options, Sbmt::KafkaConsumer::Types::ConfigAttrs.default({}.freeze)
+
+  # defaults are rdkafka's
+  # see https://github.com/confluentinc/librdkafka/blob/master/CONFIGURATION.md
+  attribute :heartbeat_timeout, Sbmt::KafkaConsumer::Types::Coercible::Integer.optional.default(5)
+  attribute :session_timeout, Sbmt::KafkaConsumer::Types::Coercible::Integer.optional.default(30)
+  attribute :reconnect_timeout, Sbmt::KafkaConsumer::Types::Coercible::Integer.optional.default(3)
+  attribute :connect_timeout, Sbmt::KafkaConsumer::Types::Coercible::Integer.optional.default(5)
+  attribute :socket_timeout, Sbmt::KafkaConsumer::Types::Coercible::Integer.optional.default(30)
+
+  attribute :kafka_options, Sbmt::KafkaConsumer::Types::ConfigAttrs.optional.default({}.freeze)
 
   def to_kafka_options
-    # servers takes precedence over kafka_options
-    kafka_options
-      .merge("bootstrap.servers": servers)
-      .symbolize_keys
+    # root options take precedence over kafka_options' ones
+    kafka_options.merge(
+      "bootstrap.servers": servers,
+      "heartbeat.interval.ms": heartbeat_timeout * 1_000,
+      "session.timeout.ms": session_timeout * 1_000,
+      "reconnect.backoff.max.ms": reconnect_timeout * 1_000,
+      "socket.connection.setup.timeout.ms": connect_timeout * 1_000,
+      "socket.timeout.ms": socket_timeout * 1_000
+    ).symbolize_keys
   end
 end